refactor shell interface into its own file
